Moved extensions of SettingsVC in a single file
